LayoutTests/imported/w3c/web-platform-tests/dom/events/event-disabled-dynamic-expected.txt



PASS disabled is honored properly in presence of dynamic changes
